fix IDX page tests
