+++ src/app/pages/IdxPage/testHelpers/index.jsx
@@ -7,7 +7,7 @@ import { service as ukrainianConfig } from '#lib/config/services/ukrainian';
 import { service as persianConfig } from '#lib/config/services/persian';
 import { ServiceContext } from '#contexts/ServiceContext';
 import { UserContextProvider } from '#contexts/UserContext';
-import IdxPage from '#pages/IdxPage';
+import IdxPage from '#app/pages/IdxPage/IdxPage';
 
 const radioServiceOverride = 'dari';
